+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/state/messages/MessagesState.kt
@@ -16,11 +16,8 @@
 
 package io.getstream.chat.android.compose.state.messages
 
-import androidx.compose.ui.unit.IntSize
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.compose.state.messages.list.MessageListItemState
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 
 /**
  * UI representation of the Conversation/Messages screen. Holds all the data required to show messages.
@@ -37,7 +34,6 @@ import kotlinx.coroutines.flow.StateFlow
  * @param startOfMessages If we're ate the start of messages (to stop pagination).
  * @param isLoadingMoreOldMessages If we're loading older messages.
  * @param isLoadingMoreNewMessages If we're loading newer messages.
- * @param focusedMessageOffsetState The offset needed to center a selected item in the list.
  */
 public data class MessagesState(
     val isLoading: Boolean = true,
@@ -52,29 +48,4 @@ public data class MessagesState(
     val startOfMessages: Boolean = false,
     val isLoadingMoreOldMessages: Boolean = false,
     val isLoadingMoreNewMessages: Boolean = false,
-    private val focusedMessageOffsetState: MutableStateFlow<Int?> = MutableStateFlow(null)
-) {
-
-    /**
-     * The offset the list needs to apply so that the focused item is centered inside the screen.
-     */
-    public val focusedMessageOffset: StateFlow<Int?> = focusedMessageOffsetState
-
-    /**
-     * Calculates the message offset needed for the message to center inside the list on scroll.
-     *
-     * @param parentSize The size of the list which contains the message.
-     * @param focusedMessageSize The size of the message item we wish to bring to the center and focus.
-     */
-    public fun calculateMessageOffset(parentSize: IntSize, focusedMessageSize: IntSize) {
-        if (parentSize.height == 0 || focusedMessageSize.height == 0) return
-
-        val sizeDiff = parentSize.height - focusedMessageSize.height
-        val offset = if (sizeDiff > 0) {
-            -sizeDiff / 2
-        } else {
-            -sizeDiff
-        }
-        if (offset != focusedMessageOffsetState.value) focusedMessageOffsetState.value = offset
-    }
-}
+)